git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rebmix@207 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/rebmix/src/base.cpp
@@ -166,7 +166,7 @@ int Digamma(FLOAT y, FLOAT *Psi)
                     Error = 1; goto E0;
                 }
 
-                aug = sgn * (cos(z) / sin(z) * (FLOAT)4.0);
+                aug = sgn * ((FLOAT)cos(z) / (FLOAT)sin(z) * (FLOAT)4.0);
             }
             else {
                 aug = sgn * ((FLOAT)sin(z) / (FLOAT)cos(z) * (FLOAT)4.0);
